Move GcAdpcm coefficient code to it's own file
