+++ src/VGAudio.Tests/Formats/GcAdpcm/GcAdpcmAlignmentTests.cs
@@ -70,7 +70,7 @@ namespace VGAudio.Tests.Formats.GcAdpcm
         {
             int loopEnd = sineCycles * 4 * SamplesPerFrame + loopStart;
             var pcm = GenerateSineWave(GetNextMultiple(loopEnd, SamplesPerFrame), 1, SamplesPerFrame * 4);
-            var coefs = GcAdpcmEncoder.DspCorrelateCoefs(pcm);
+            var coefs = GcAdpcmCoefficient.CalculateCoefficients(pcm);
             var adpcm = GcAdpcmEncoder.EncodeAdpcm(pcm, coefs);
             var alignment = new GcAdpcmAlignment(multiple, loopStart, loopEnd, adpcm, coefs);
             var pcmAligned = GcAdpcmDecoder.Decode(alignment.AdpcmAligned, coefs, alignment.SampleCountAligned);
@@ -98,7 +98,7 @@ namespace VGAudio.Tests.Formats.GcAdpcm
         {
             int loopEnd = sineCycles * 4 * SamplesPerFrame + loopStart;
             var pcm = GenerateSineWave(GetNextMultiple(loopEnd, SamplesPerFrame), 1, SamplesPerFrame * 4);
-            var coefs = GcAdpcmEncoder.DspCorrelateCoefs(pcm);
+            var coefs = GcAdpcmCoefficient.CalculateCoefficients(pcm);
             var adpcm = GcAdpcmEncoder.EncodeAdpcm(pcm, coefs);
             var alignment = new GcAdpcmAlignment(multiple, loopStart, loopEnd, adpcm, coefs);
             var pcmAligned = alignment.PcmAligned;
